pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "ssdd-calculator"
version = "0.1.0"
authors = [
  { name="Tobias Diaz", email="tobias.diaz@uclm.es" },
  { name="Jose Luis Segura", email="joseluis.segura@uclm.es"},
]
description = "Template for Distributed Systems laboratory"
requires-python = ">=3.10"
dependencies = [
    'zeroc-ice',
    'pydantic',
    'pytest',
    'confluent-kafka',
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"

[project.scripts]
ssdd-calculator = "calculator.command_handlers:calculator"
ssdd-kafka = "calculator.kafka:main"
